# core/vm: remove pointers from JumpTable