no reason for them to be pointers since they are all non-nil. Remove unused "undefined" bit as well.