@omerfirmak 2025-07-09
rebased and done a quick benchmark

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm/runtime
cpu: AMD Ryzen 7 5700U with Radeon Graphics         
                                       │ ptr-jit.txt │           no-ptr-jit.txt            │
                                       │   sec/op    │   sec/op     vs base                │
SimpleLoop/staticcall-identity-100M-16   294.0m ± 0%   277.2m ± 0%   -5.72% (p=0.000 n=10)
SimpleLoop/call-identity-100M-16         338.5m ± 0%   332.4m ± 1%   -1.79% (p=0.000 n=10)
SimpleLoop/loop-100M-16                  335.0m ± 1%   289.5m ± 1%  -13.58% (p=0.000 n=10)
SimpleLoop/loop2-100M-16                 287.2m ± 1%   285.4m ± 0%        ~ (p=0.075 n=10)
SimpleLoop/call-nonexist-100M-16         663.5m ± 0%   653.3m ± 0%   -1.54% (p=0.000 n=10)
SimpleLoop/call-EOA-100M-16              271.0m ± 0%   267.9m ± 1%   -1.15% (p=0.001 n=10)
SimpleLoop/call-reverting-100M-16        532.6m ± 0%   541.6m ± 1%   +1.67% (p=0.000 n=10)
geomean                                  368.2m        355.8m        -3.37%

                                       │ ptr-jit.txt  │           no-ptr-jit.txt            │
                                       │     B/op     │     B/op      vs base               │
SimpleLoop/staticcall-identity-100M-16   67.95Mi ± 0%   67.95Mi ± 0%       ~ (p=0.950 n=10)
SimpleLoop/call-identity-100M-16         66.58Mi ± 0%   66.58Mi ± 0%  +0.00% (p=0.000 n=10)
SimpleLoop/loop-100M-16                  5.458Ki ± 8%   5.273Ki ± 4%       ~ (p=0.089 n=10)
SimpleLoop/loop2-100M-16                 5.273Ki ± 4%   5.273Ki ± 4%       ~ (p=1.000 n=10)
SimpleLoop/call-nonexist-100M-16         119.6Mi ± 0%   119.6Mi ± 0%       ~ (p=0.052 n=10)
SimpleLoop/call-EOA-100M-16              22.79Mi ± 0%   22.79Mi ± 0%       ~ (p=0.442 n=10)
SimpleLoop/call-reverting-100M-16        163.8Mi ± 0%   163.8Mi ± 0%  +0.02% (p=0.000 n=10)
geomean                                  4.761Mi        4.737Mi       -0.49%

                                       │ ptr-jit.txt │           no-ptr-jit.txt           │
                                       │  allocs/op  │  allocs/op   vs base               │
SimpleLoop/staticcall-identity-100M-16   2.740M ± 0%   2.740M ± 0%       ~ (p=0.734 n=10)
SimpleLoop/call-identity-100M-16         2.685M ± 0%   2.685M ± 0%       ~ (p=0.211 n=10)
SimpleLoop/loop-100M-16                   41.00 ± 2%    41.00 ± 2%       ~ (p=0.303 n=10)
SimpleLoop/loop2-100M-16                  41.00 ± 2%    41.00 ± 2%       ~ (p=1.000 n=10)
SimpleLoop/call-nonexist-100M-16         2.985M ± 0%   2.985M ± 0%       ~ (p=0.068 n=10)
SimpleLoop/call-EOA-100M-16              746.3k ± 0%   746.3k ± 0%       ~ (p=0.312 n=10)
SimpleLoop/call-reverting-100M-16        2.858M ± 0%   2.858M ± 0%  +0.00% (p=0.000 n=10)
geomean                                  96.64k        96.64k       +0.00%
```